+++ lntest/itest/lnd_channel_backup_test.go
@@ -1,5 +1,3 @@
-// +build rpctest
-
 package itest
 
 import (
