+++ lntest/itest/lnd_forward_interceptor_test.go
@@ -1,5 +1,3 @@
-// +build rpctest
-
 package itest
 
 import (
