+++ lntest/itest/lnd_multi-hop_htlc_local_timeout_test.go
@@ -1,5 +1,3 @@
-// +build rpctest
-
 package itest
 
 import (
