+++ lntest/itest/lnd_multi-hop_htlc_remote_chain_claim_test.go
@@ -1,5 +1,3 @@
-// +build rpctest
-
 package itest
 
 import (
