+++ lntest/itest/lnd_multi-hop_remote_force_close_on_chain_htlc_timeout_test.go
@@ -1,5 +1,3 @@
-// +build rpctest
-
 package itest
 
 import (
