+++ lntest/itest/lnd_network_test.go
@@ -1,5 +1,3 @@
-// +build rpctest
-
 package itest
 
 import (
