+++ lntest/itest/lnd_send_multi_path_payment_test.go
@@ -1,5 +1,3 @@
-// +build rpctest
-
 package itest
 
 import (
