+++ lntest/itest/lnd_single_hop_invoice_test.go
@@ -1,5 +1,3 @@
-// +build rpctest
-
 package itest
 
 import (
