+++ lntest/itest/lnd_test_list_off_test.go
@@ -1,0 +1,5 @@
+// +build !rpctest
+
+package itest
+
+var testsCases = []*testCase{}
