+++ lntest/itest/lnd_wumbo_channels_test.go
@@ -1,5 +1,3 @@
-// +build rpctest
-
 package itest
 
 import (
